requirements.txt
yfinance>=0.2.36
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.3
python-dotenv>=1.0.0
httpx>=0.26.0
pyyaml>=6.0
